implement(map historical):Designing a web page to display 2D airsoft maps
Thanks to this commit, the client interface is now able to show images of maps that was scanned by our robot.
An icon button is  there to give the opportunity to open maps in a new tab.
The image list was implemented with the help of the Material UI <ImageList> component.