Implemented user registration using Firebase Authentication
This commit introduces the ability to create an account on the RoomMapping web application through Firebase Authentication.
During the registration process, email and password formats are validated.
Additionally, the system checks if an account with the provided email already exists.
If all the input fields in the form are correct, the account is created in the Firebase project.
A brief animation notifies the user of the successful account creation, and the user is then redirected to the app's login page.

This feature enhances user onboarding and provides a seamless registration experience.