Implemented user account creation and authentication
This commit enables the creation of an account and the authentication of a specific account managed by Firebase.
Currently, the app does not retain information about the connected user, and there is no concept of user sessions.

This functionality establishes the foundation for user management within the RoomMapping application.
Future enhancements will focus on incorporating user session management for a more personalized and secure user experience.