Us8 (#39)
This pull request will merge the code for the feature described in US8 : being able to make a map accessible to players via the use of a qrcode. This is only functional for the static version for now but it should be setup for easy implementation with the dynamic version of the map.

It has been implemented with two popup components allowing changes to the parameters as well as visualisation of the qrcode in the frontend and a qrcode generating route in the backend.

* backend

Added r_qrcode.js, cl_qrcode.js to handle the creation/registration, modification and deletion of a maps publishing.

* Frontend code for publishing a map

Added two popus to Maplist.js to let the user publish a map and subsequently view it's qrcode or change the way it's published

* Testing set up

set up the testing for us8. didn't get any further due to fetch not working, merging anyways due to time constraints